Too much trouble
I'll remove the possibility to make quit saves. There are too many issues to deal with and I'll have to rewrite too much code from scratch to come with a truly satisfying fix. With bosses respawning and stuff this is also the safer way to go and now I can also close #53 as that issue is also voided in the process (except for debug saves, but they are not available to the player anyway).